Added visual indicator for audio
AudioTimelinePanel is a new file with everything related to the visualizer exists.
Within SlideshowCreator, two new functions: updateAudioTimeline and calculateTotalSlideshowDuration.
The updateAudioTimeline one is used to do exactly it's name. The other function is temporary until we enable for the user to select duration per image. For now it assumes 5 seconds per image, and if there are less than 3 images it will default to 15 seconds for testing reasons.