First attempt to show more accurate audio bar
This is a little of a weird one but follow along.

So since the TimelinePanel doesn't keep track of slide length, I have to calculate that in audioTimelinePanel. Additionally, both panels are not synced together at all, so to show audio length to slide length, unorthodox methods are necessary.

What I've opted to attempt in this case is that I do the slide calculations and project them in the audioTimelinePanel as alternating background colors. So if you see a light grey background section, that's the length on slide 1's length. The dark grey background next is slide 2's length (inluding transition duration). Then for however many more slides, the background colors continue to alternate.

Then the class blue audio file thing I've previously used is now mapped on top of these alternating background colors representing individual slide lengths. Thus the user can somewhat gauge at what point within a slide and audio shall begin/end.

NOTE: From my tests, it's not quite accurate so my math is probably wrong from what I'm doing, however it is vaguely close. This is a step in the right direction so I encourage playing around with it more.